pyproject.toml
[project]
name = "quotepedia"
version = "0.0.0"
dependencies = [
  "fastapi[all]>=0.112.0",
  "fastapi-mail>=1.4.1",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "psycopg>=3.2.1",
  "psycopg-binary>=3.2.1",
  "psycopg2>=2.9.9",
  "psycopg2-binary>=2.9.9",
  "pyheck>=0.1.5",
  "pyjwt>=2.9.0",
  "passlib>=1.7.4",
  "bcrypt==4.0.1",
  "redis>=5.0.8",
  "pillow>=10.4.0",
]
requires-python = "==3.12.*"


[tool.pdm]
distribution = false


[tool.pdm.dev-dependencies]
dev = [
  "ruff>=0.5.7",
  "babel>=2.16.0",
]


[tool.pdm.scripts]
"dev" = "uvicorn src.app:app --reload"
"start" = "uvicorn src.app:app --host 0.0.0.0 --port 8000"
"check" = "ruff check"
"format" = "ruff format"
"i18n:update" = "pybabel update -w 120 -D messages -i 'src\\i18n\\locales\\messages.pot' -d 'src\\i18n\\locales'"
"i18n:extract" = "pybabel extract -w 120 -F babel.ini -o 'src\\i18n\\locales\\messages.pot' ."
"i18n:compile" = "pybabel compile -f -D messages -d 'src\\i18n\\locales' --statistics"


[tool.ruff]
line-length = 120


[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # Pyflakes
  "I", # isort
]
